refactor: remove bound-signal
